backend/requirements.txt
asgiref==3.2.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==2.2.16
django-extensions==3.1.5
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
flake8==4.0.1
gunicorn==20.0.4
idna==3.3
iniconfig==1.1.1
install==1.3.4
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mccabe==0.6.1
oauthlib==3.2.0
packaging==21.2
Pillow==10.1.0
pluggy==0.13.1
psycopg2-binary==2.9.9
py==1.10.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==2.4.7
pytest==7.0.1
pytest-pythonpath==0.7.3
python-dotenv==0.13.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.1
sentry-sdk==1.5.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
style==1.1.0
toml==0.10.2
tomli==2.0.1
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.7
